Rename old merkle to merkle_normal
