+++ tests/eth2/merkle-utils/test_merkle_trees.py
@@ -8,7 +8,7 @@ from eth2.beacon._utils.hash import (
     hash_eth2,
 )
 
-from eth2._utils.merkle import (
+from eth2._utils.merkle_normal import (
     get_merkle_root_from_items,
     calc_merkle_tree,
     get_root,
